requirements.txt
discord.py
python-dotenv
SpeechRecognition
edge-tts
pydub
